11.3.2 upgrade

+++ docroot/core/modules/views/src/DisplayPluginCollection.php
@@ -5,11 +5,15 @@ namespace Drupal\views;
 use Drupal\Component\Plugin\Exception\PluginException;
 use Drupal\Component\Plugin\PluginManagerInterface;
 use Drupal\Core\Plugin\DefaultLazyPluginCollection;
+use Drupal\Core\StringTranslation\StringTranslationTrait;
+use Drupal\views\Plugin\views\display\DisplayPluginInterface;
 
 /**
  * A class which wraps the displays of a view so you can lazy-initialize them.
  */
 class DisplayPluginCollection extends DefaultLazyPluginCollection {
+
+  use StringTranslationTrait;
 
   /**
    * Stores a reference to the view which has this displays attached.
@@ -49,6 +53,7 @@ class DisplayPluginCollection extends DefaultLazyPluginCollection {
    * {@inheritdoc}
    *
    * @return \Drupal\views\Plugin\views\display\DisplayPluginBase
+   *   The display plugin.
    */
   public function &get($instance_id) {
     return parent::get($instance_id);
@@ -59,7 +64,9 @@ class DisplayPluginCollection extends DefaultLazyPluginCollection {
    */
   public function clear() {
     foreach (array_filter($this->pluginInstances) as $display) {
-      $display->destroy();
+      if ($display instanceof DisplayPluginInterface) {
+        $display->destroy();
+      }
     }
 
     parent::clear();
@@ -80,7 +87,7 @@ class DisplayPluginCollection extends DefaultLazyPluginCollection {
     // display plugin isn't found.
     catch (PluginException $e) {
       $message = $e->getMessage();
-      \Drupal::messenger()->addWarning(t('@message', ['@message' => $message]));
+      \Drupal::messenger()->addWarning($this->t('@message', ['@message' => $message]));
     }
 
     // If no plugin instance has been created, return NULL.
